[IRGen][DebugInfo] Store demangled names for all function entities in CodeView
Windows crash dumps are difficult to understand, because their backtraces
are full of mangled names. macOS appears to apply Swift demangling during
symbolication. This is not yet possible on Windows. Instead we can write
demangled names to PDBs for all function-like entities at compile-time.

This change can cause an increase of up to 10% in PDB file size. It does
not affect object file size. The improvement for the quality of crash
dumps is significant.